requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt[crypto]>=5.3.0
PyJWT>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.1
